# openapi.yaml
openapi: 3.0.3
info:
  title: AI-Powered HRMS API
  description: Next-generation Human Resource Management System with AI-powered resume screening and conversational features
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3001/api
    description: Local development server
  - url: https://your-production-url.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employee_code:
          type: string
        full_name:
          type: string
        position:
          type: string
        department_id:
          type: integer
        department_name:
          type: string
        employment_type:
          type: string
          enum: [full-time, part-time, contract, intern]
        hire_date:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, terminated, on-leave]
    JobPosting:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        responsibilities:
          type: string
        department_id:
          type: integer
        employment_type:
          type: string
        experience_level:
          type: string
          enum: [entry, mid, senior, lead]
        required_skills:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, closed, on-hold]
        published_at:
          type: string
          format: date-time
    Application:
      type: object
      properties:
        id:
          type: integer
        job_id:
          type: integer
        candidate_name:
          type: string
        candidate_email:
          type: string
          format: email
        candidate_phone:
          type: string
        years_of_experience:
          type: number
        screening_score:
          type: number
          format: float
        status:
          type: string
          enum: [received, screening, screened, shortlisted, interview, rejected, offer, hired]
        applied_at:
          type: string
          format: date-time
paths:
...
  /employees:
    get:
      tags:
        - Employees
      summary: List employees
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            default: active
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items:
                      $ref: '#/components/schemas/Employee'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer

    post:
      tags:
        - Employees
      summary: Create new employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - employeeCode
                - fullName
                - hireDate
              properties:
                employeeCode:
                  type: string
                fullName:
                  type: string
                position:
                  type: string
                departmentId:
                  type: integer
                hireDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  employee:
                    $ref: '#/components/schemas/Employee'
  /jobs:
    get:
      tags:
        - Jobs
      summary: List job postings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            default: published
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobPosting'
                  pagination:
                    type: object

    post:
      tags:
        - Jobs
      summary: Create new job posting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
              properties:
                title:
                  type: string
                description:
                  type: string
                requirements:
                  type: string
                requiredSkills:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  job:
                    $ref: '#/components/schemas/JobPosting'
  /jobs/{id}:
    get:
      tags:
        - Jobs
      summary: Get job posting details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                type: object
                properties:
                  job:
                    $ref: '#/components/schemas/JobPosting'
...
  /applications:
    get:
      tags:
        - Applications
      summary: List applications
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: minScore
          in: query
          schema:
            type: number
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Application'
                  pagination:
                    type: object
...
  /chat:
    post:
      tags:
        - Chat
      summary: Chat with AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                conversationId:
                  type: integer
                applicationId:
                  type: integer
                message:
                  type: string
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversationId:
                    type: integer
                  response:
                    type: string
                  provider:
                    type: string